+++ src/main/java/net/fabricmc/mappingio/adapter/ForwardingMappingVisitor.java
@@ -27,6 +27,11 @@ import net.fabricmc.mappingio.MappedElementKind;
 import net.fabricmc.mappingio.MappingFlag;
 import net.fabricmc.mappingio.MappingVisitor;
 
+/**
+ * A mapping visitor that forwards all visit calls to another {@link MappingVisitor}.
+ *
+ * <p>Subclasses should override the visit methods they want to intercept.
+ */
 public abstract class ForwardingMappingVisitor implements MappingVisitor {
 	protected ForwardingMappingVisitor(MappingVisitor next) {
 		Objects.requireNonNull(next, "null next");