+++ src/main/java/net/fabricmc/mappingio/format/tiny/Tiny1FileWriter.java
@@ -28,7 +28,11 @@ import org.jetbrains.annotations.Nullable;
 import net.fabricmc.mappingio.MappedElementKind;
 import net.fabricmc.mappingio.MappingFlag;
 import net.fabricmc.mappingio.MappingWriter;
+import net.fabricmc.mappingio.format.MappingFormat;
 
+/**
+ * {@linkplain MappingFormat#TINY_1 Tiny v1 file} writer.
+ */
 public final class Tiny1FileWriter implements MappingWriter {
 	public Tiny1FileWriter(Writer writer) {
 		this.writer = writer;