+++ src/test/java/net/fabricmc/mappingio/TestHelper.java
@@ -28,7 +28,7 @@ import org.cadixdev.lorenz.io.MappingFormats;
 import org.jetbrains.annotations.Nullable;
 
 import net.fabricmc.mappingio.format.MappingFormat;
-import net.fabricmc.mappingio.tree.MappingTree;
+import net.fabricmc.mappingio.tree.MappingTreeView;
 import net.fabricmc.mappingio.tree.MemoryMappingTree;
 
 public final class TestHelper {
@@ -119,7 +119,7 @@ public final class TestHelper {
 		}
 	}
 
-	public static Path writeToDir(MappingTree tree, MappingFormat format, Path dir) throws IOException {
+	public static Path writeToDir(MappingTreeView tree, Path dir, MappingFormat format) throws IOException {
 		Path path = dir.resolve(getFileName(format));
 		tree.accept(MappingWriter.create(path, format));
 		return path;
@@ -257,7 +257,7 @@ public final class TestHelper {
 	}
 
 	private static void visitMethodArg(MemoryMappingTree tree, int... dstNs) {
-		tree.visitMethodArg(counter.getAndIncrement(), counter.getAndIncrement(), nameGen.src(argKind));
+		tree.visitMethodArg(nameGen.getCounter().getAndIncrement(), nameGen.getCounter().getAndIncrement(), nameGen.src(argKind));
 
 		for (int ns : dstNs) {
 			tree.visitDstName(argKind, ns, nameGen.dst(argKind, ns));
@@ -265,7 +265,8 @@ public final class TestHelper {
 	}
 
 	private static void visitMethodVar(MemoryMappingTree tree, int... dstNs) {
-		tree.visitMethodVar(counter.get(), counter.get(), counter.getAndIncrement(), counter.getAndIncrement(), nameGen.src(varKind));
+		tree.visitMethodVar(nameGen.getCounter().get(), nameGen.getCounter().get(),
+				nameGen.getCounter().getAndIncrement(), nameGen.getCounter().getAndIncrement(), nameGen.src(varKind));
 
 		for (int ns : dstNs) {
 			tree.visitDstName(varKind, ns, nameGen.dst(varKind, ns));
@@ -286,6 +287,7 @@ public final class TestHelper {
 			argNum.get().set(0);
 			varNum.get().set(0);
 			nsNum.get().set(0);
+			counter.get().set(0);
 		}
 
 		private void resetNsNum() {
@@ -353,6 +355,10 @@ public final class TestHelper {
 			return sb.toString();
 		}
 
+		public AtomicInteger getCounter() {
+			return counter.get();
+		}
+
 		private AtomicInteger getCounter(MappedElementKind kind) {
 			switch (kind) {
 			case CLASS:
@@ -401,9 +407,17 @@ public final class TestHelper {
 		private ThreadLocal<AtomicInteger> argNum = ThreadLocal.withInitial(() -> new AtomicInteger());
 		private ThreadLocal<AtomicInteger> varNum = ThreadLocal.withInitial(() -> new AtomicInteger());
 		private ThreadLocal<AtomicInteger> nsNum = ThreadLocal.withInitial(() -> new AtomicInteger());
+		private ThreadLocal<AtomicInteger> counter = ThreadLocal.withInitial(() -> new AtomicInteger());
 	}
 
 	public static class MappingDirs {
+		@Nullable
+		public static MemoryMappingTree getCorrespondingTree(Path dir) {
+			if (dir.equals(VALID)) return createTestTree();
+			if (dir.equals(VALID_WITH_HOLES)) return createTestTreeWithHoles();
+			return null;
+		}
+
 		public static final Path DETECTION = getResource("/detection/");
 		public static final Path VALID = getResource("/read/valid/");
 		public static final Path VALID_WITH_HOLES = getResource("/read/valid-with-holes/");
@@ -413,7 +427,6 @@ public final class TestHelper {
 	private static final String mthDesc = "()I";
 	private static final String comment = "This is a comment";
 	private static final NameGen nameGen = new NameGen();
-	private static final AtomicInteger counter = new AtomicInteger();
 	private static final MappedElementKind clsKind = MappedElementKind.CLASS;
 	private static final MappedElementKind fldKind = MappedElementKind.FIELD;
 	private static final MappedElementKind mthKind = MappedElementKind.METHOD;